+++ src/algorithm/sort/leetcode/reversePairs.ts
@@ -1,40 +1,34 @@
-interface Item {
-  value: number;
-  index: number;
-}
-
-// https://leetcode-cn.com/problems/count-of-smaller-numbers-after-self/
-// 315
-function countSmaller(nums: number[]): number[] {
-  const count: number[] = Array(nums.length)
-    .fill(0);
-
-  function mergeSort(originalArray: Item[], left = 0, right = originalArray.length - 1): void {
+// 类似 https://leetcode-cn.com/problems/shu-zu-zhong-de-ni-xu-dui-lcof/
+// 剑指 Offer 51. 数组中的逆序对
+export function reversePairs(nums: number[]): number {
+  let count = 0;
+  function mergeSort(originalArray: number[], left = 0, right = originalArray.length - 1): number[] {
     if (left >= right) {
-      return;
+      return originalArray;
     }
 
     const middleIndex = (left + right) >>> 1;
     mergeSort(originalArray, left, middleIndex);
     mergeSort(originalArray, middleIndex + 1, right);
     mergeSortedArrays(originalArray, left, right);
+    return originalArray;
   }
 
-  function mergeSortedArrays(originalArray: Item[], left: number, right: number): void {
-    const result: Item[] = Array(right - left + 1);
+  function mergeSortedArrays(originalArray: number[], left: number, right: number): void {
+    const result: number[] = Array(right - left + 1);
     const middleIndex = (left + right) >>> 1;
     let i = left;
     let j = middleIndex + 1;
     let k = 0;
 
     while (i <= middleIndex && j <= right) {
-      if (originalArray[j].value < originalArray[i].value) {
+      if (originalArray[j] < originalArray[i]) {
         result[k] = originalArray[j];
         k++;
         j++;
       } else {
         result[k] = originalArray[i];
-        count[originalArray[i].index] += (j - 1) - (middleIndex + 1) + 1;
+        count += (j - 1) - (middleIndex + 1) + 1;
         k++;
         i++;
       }
@@ -42,7 +36,7 @@ function countSmaller(nums: number[]): number[] {
 
     while (i <= middleIndex) {
       result[k] = originalArray[i];
-      count[originalArray[i].index] += (j - 1) - (middleIndex + 1) + 1;
+      count += (j - 1) - (middleIndex + 1) + 1;
       k++;
       i++;
     }
@@ -58,9 +52,6 @@ function countSmaller(nums: number[]): number[] {
     }
   }
 
-  mergeSort(nums.map((item, index) => ({
-    value: item,
-    index,
-  })));
+  mergeSort(nums);
   return count;
 }