+++ src/dataStructures/queue/circularQueue.ts
@@ -5,52 +5,51 @@ export class CircularQueue {
 
   private readonly capacity: number;
 
-  private head: number;
+  private head = 0;
 
-  private tail: number;
+  private tail = 0;
 
   constructor(k: number) {
     this.capacity = k + 1;
     this.data = Array(this.capacity)
       .fill(null);
-    this.head = 0;
     this.tail = 0;
   }
 
   public enQueue(value: number): boolean {
     if (this.isFull()) {
       return false;
-    } else {
-      this.data[this.tail] = value;
-      this.tail = (this.tail + 1) % this.capacity;
-      return true;
     }
+
+    this.data[this.tail] = value;
+    this.tail = (this.tail + 1) % this.capacity;
+    return true;
   }
 
   public deQueue(): boolean {
     if (this.isEmpty()) {
       return false;
-    } else {
-      this.data[this.head] = null;
-      this.head = (this.head + 1) % this.capacity;
-      return true;
     }
+
+    this.data[this.head] = null;
+    this.head = (this.head + 1) % this.capacity;
+    return true;
   }
 
   public Front(): number {
     if (this.isEmpty()) {
       return -1;
-    } else {
-      return this.data[this.head];
     }
+
+    return this.data[this.head];
   }
 
   public Rear(): number {
     if (this.isEmpty()) {
       return -1;
-    } else {
-      return this.data[(this.tail - 1 + this.capacity) % this.capacity];
     }
+
+    return this.data[(this.tail - 1 + this.capacity) % this.capacity];
   }
 
   public isEmpty(): boolean {